@@ -1,33 +1,42 @@
-# encoding: UTF-8
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'spree_slider'
-  s.version     = '3.3.0'
+  s.version     = '4.0.0'
   s.summary     = 'Spree Slider extension'
   s.description = 'Adds a slider to the homepage'
-  s.required_ruby_version = '>= 1.8.7'
+  s.required_ruby_version = '>= 2.5.0'
 
   s.author            = 'Giuseppe Privitera'
   s.email             = 'priviterag@gmail.com'
-  s.homepage          = 'https://github.com/priviterag/spree_slider'
+  s.homepage          = 'https://github.com/spree-contrib/spree_slider'
 
   s.files         = `git ls-files`.split("\n")
   s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  spree_version = '>= 3.1.0', '< 4.0'
-  s.add_dependency 'spree_core', spree_version
+  spree_version = '>= 4.0', '< 5.0'
+  s.add_dependency 'active_storage_validations', '~> 0.8.4'
   s.add_dependency 'spree_backend', spree_version
+  s.add_dependency 'spree_core', spree_version
+  s.add_dependency 'spree_extension'
 
   s.add_development_dependency 'appraisal'
-  s.add_development_dependency 'factory_girl'
-  s.add_development_dependency 'factory_girl_rails'
+  s.add_development_dependency 'awesome_print'
+  s.add_development_dependency 'byebug'
+  s.add_development_dependency 'capybara'
+  s.add_development_dependency 'capybara-screenshot'
+  s.add_development_dependency 'coffee-rails'
+  s.add_development_dependency 'database_cleaner'
+  s.add_development_dependency 'factory_bot', '~> 4.7'
   s.add_development_dependency 'ffaker'
-  s.add_development_dependency 'guard-rspec'
   s.add_development_dependency 'mysql2'
   s.add_development_dependency 'pg'
-  s.add_development_dependency 'rspec-rails'
-  s.add_development_dependency 'shoulda-matchers'
+  s.add_development_dependency 'pry'
+  s.add_development_dependency 'rspec-mocks'
+  s.add_development_dependency 'rspec-rails', '~> 4.0.0.beta2'
+  s.add_development_dependency 'sass-rails'
+  s.add_development_dependency 'simplecov'
   s.add_development_dependency 'sqlite3'
+  s.add_development_dependency 'webdrivers', '~> 4.0.0'
 end
