@@ -1,8 +1,14 @@
-Spree::Product.class_eval do
-  has_one :slide, dependent: :destroy
-  after_update :destroy_slide_if_deleted
+module Spree
+  module ProductDecorator
+    def self.prepended(base)
+      base.has_one :slide
+      base.after_destroy :destroy_slide_if_deleted
+    end
 
-  def destroy_slide_if_deleted
-    slide.update_attributes(published: false) if slide && deleted_at
+    def destroy_slide_if_deleted
+      slide.update(published: false) if slide && deleted_at
+    end
   end
 end
+
+::Spree::Product.prepend(Spree::ProductDecorator)
