@@ -4,19 +4,17 @@ module SpreeSlider
     isolate_namespace Spree
     engine_name 'spree_slider'
 
-    config.autoload_paths += %W(#{config.root}/lib)
-
     # use rspec for tests
     config.generators do |g|
       g.test_framework :rspec
     end
 
     def self.activate
-      Dir.glob(File.join(File.dirname(__FILE__), "../../app/**/*_decorator*.rb")) do |c|
+      Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/*_decorator*.rb')) do |c|
         Rails.configuration.cache_classes ? require(c) : load(c)
       end
     end
 
-    config.to_prepare &method(:activate).to_proc
+    config.to_prepare(&method(:activate).to_proc)
   end
 end
