@@ -1,21 +1,73 @@
-ENV["RAILS_ENV"] ||= 'test'
-require File.expand_path("../dummy/config/environment", __FILE__)
+# Run Coverage report
+require 'simplecov'
+
+SimpleCov.start do
+  add_filter 'spec/dummy'
+  add_filter 'spec/support'
+  add_group 'Controllers', 'app/controllers'
+  add_group 'Helpers', 'app/helpers'
+  add_group 'Mailers', 'app/mailers'
+  add_group 'Models', 'app/models'
+  add_group 'Views', 'app/views'
+  add_group 'Libraries', 'lib/spree'
+end
+
+# Configure Rails Environment
+ENV['RAILS_ENV'] = 'test'
+
+require File.expand_path('dummy/config/environment.rb', __dir__)
+
 require 'rspec/rails'
+require 'rspec/mocks'
+require 'ffaker'
+require 'pry'
+require 'active_storage_validations'
 
-# Requires supporting files with custom matchers and macros, etc,
-# in ./support/ and its subdirectories.
-Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
+# Requires supporting ruby files with custom matchers and macros, etc,
+# in spec/support/ and its subdirectories.
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
 
-require 'spree/core/testing_support/factories'
-require 'spree/core/testing_support/env'
-require 'spree/core/testing_support/controller_requests'
-require 'spree/core/url_helpers'
+# Requires factories and other useful helpers defined in spree_core.
+require 'spree/testing_support/authorization_helpers'
+require 'spree/testing_support/capybara_ext'
+require 'spree/testing_support/controller_requests'
+require 'spree/testing_support/factories'
+require 'spree/testing_support/url_helpers'
+
+# Requires factories defined in lib/spree_slider/factories.rb
+require 'spree_slider/factories'
 
 RSpec.configure do |config|
-  config.mock_with :rspec
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
-  config.use_transactional_fixtures = true
+  # Infer an example group's spec type from the file location.
+  config.infer_spec_type_from_file_location!
 
-  config.include Spree::Core::TestingSupport::ControllerRequests
-  config.include Devise::TestHelpers, :type => :controller
+  # == URL Helpers
+  #
+  # Allows access to Spree's routes in specs:
+  #
+  # visit spree.admin_path
+  # current_path.should eql(spree.products_path)
+  config.include Spree::TestingSupport::UrlHelpers
+
+  # == Mock Framework
+  #
+  # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
+  #
+  # config.mock_with :mocha
+  # config.mock_with :flexmock
+  # config.mock_with :rr
+  config.mock_with :rspec
+  config.color = true
+  config.default_formatter = 'doc'
+
+  # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
+  config.fixture_path = "#{::Rails.root}/spec/fixtures"
+
+  # Capybara javascript drivers require transactional fixtures set to false, and we use DatabaseCleaner
+  # to cleanup after each test instead.  Without transactional fixtures set to false the records created
+  # to setup a test will be unavailable to the browser, which runs under a separate server instance.
+  config.use_transactional_fixtures = false
+
+  config.fail_fast = ENV['FAIL_FAST'] || false
+  config.order = 'random'
 end
